requirements.txt
# Core Dependencies
torch>=2.0.1
transformers>=4.30.2
datasets>=2.14.0

# NLP Libraries
nltk>=3.8.1
rouge-score>=0.1.2
sentencepiece>=0.1.99

# Scientific Computing
numpy>=1.24.3
pandas>=2.0.3
scipy>=1.11.1
scikit-learn>=1.3.0

# Web Framework
flask>=2.3.0
gunicorn>=21.2.0

# File Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Visualization
networkx>=3.1
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
sphinx>=7.0.1